+++ Source/Scene/Cesium3DTileContentFactory.js
@@ -53,6 +53,15 @@ var Cesium3DTileContentFactory = {
     );
   },
   i3dm: function (tileset, tile, resource, arrayBuffer, byteOffset) {
+    if (tileset.enableModelExperimental) {
+      return ModelExperimental3DTileContent.fromI3dm(
+        tileset,
+        tile,
+        resource,
+        arrayBuffer,
+        byteOffset
+      );
+    }
     return new Instanced3DModel3DTileContent(
       tileset,
       tile,