+++ Source/Scene/ModelComponents.js
@@ -522,6 +522,16 @@ function Instances() {
    * @private
    */
   this.featureIdAttributes = [];
+
+  /**
+   * Whether the instancing transforms are applied in world space. For glTF models that
+   * use EXT_mesh_gpu_instancing, the transform is applied in object space. For i3dm files,
+   * the instance transform is in world space.
+   *
+   * @type {Boolean}
+   * @private
+   */
+  this.transformInWorldSpace = false;
 }
 
 /**